refactor: restructure UI package
