+++ src/mr_manager/ui/__init__.py
@@ -1,0 +1,3 @@
+"""UI package for the mr-manager Textual application."""
+
+from .app import MrManagerApp as MrManagerApp
