+++ src/mr_manager/ui/ui.tcss
